[NEW] CFG language for boolean operations. The new statements are:
bvar := linear_constraint
bvar := bvar
bvar := bvar <bop> bvar (<bop> is and, or, xor)
assume (bvar)
assert (bvar)
select (bvar, bvar, bvar, bvar)

The statement bvar:= bvar is not strictly necessary since it can be
simulated by e.g., bvar := bvar or false, but we think it is more
convenient like this. However, the boolean not is not directly
available but it can be simulated by bvar := bvar xor true.