Disable refinement of widening in apron domains.
We refine the result of widening in the apron domains with the result
of widening+thresholds in the interval domain. The idea is that apron
domains can benefit from thresholds without much implementation
effort. This is causing some non-termination issues. This needs more
investigation. We disable the refinement for now. We also opened
issue#20.