Return bottom instead of top if basic block not in the invariant table.
If the basic block is not in the invariant table it must be because it
was not reached by the analysis. Before, we returned top but it never
had real effect because process_pre made sure that all unreachable
blocks were in the invariant table with a bottom invariant. This was
just a waste of space/time.