Array map (map from arrays to scalars) becomes global.
This means that all abstract states share the same array map.  This
fixes some subtle bug we had but it is error-prone if multiple
analyses are run because the array map will persist from one run to
another. As an ugly solution we provide the method clear_global_state
that needs to be explicitly called.