Support to translate to linexpr and lincons elina (disabled by default).
Elina can handle both tree expressions (texpr/tcons) and linear
expressions (linexpr/lincons). The former is more general than the
latter. By default, we translate to tree expressions. This means that
Elina domains linearize internally tree expressions so that the domain
can digest the expression. We provide a pragma option USE_TREE_EXP
which is enabled by default. If this option is disabled, we might get
imprecise results if the linear expression cannot be expressed in the
domain.