Towards sound DBM domains in presence of overflow.
We added a class to represent signed 64 bits integers. Each arithmetic
operation checks for overflow.

We can plug-in this new class as DBM weights. This will make sure that
no DBM operation overflows. A current limitation is that the
implementation raises an error if overflow instead of ignoring the
operation. This is an important TODO thing.

Note that even if the DBM operations do not overflow it does not mean
that the analysis can make a wrong claim in case some program
operation overflows. For that, we implemented wrapped_numerical_domain
(wrapped_interval_domain.hpp). The idea was to use the wrapped
interval domain to correct an arbitrary numerical domain.

TODO: I should look at a simpler version where the DBM domain can
correct itself which would be probably less precise but faster.