Cleanup of the congruence domain.
Previous code assumed a predefined bitwidth so that the analysis of
bitwise operation can be more precise.  However, it assumed that *all*
variables had the same predefined bitwidth. This is not realistic for
the kind of programs we care about (e.g., C programs).

The cleanup eliminates this fixed bitwidth so the code is simpler and
cleaner but it reasons only on mathematical integers.  If needed, we
should implement separately a machine-arithmetic congruence domain
based on Stefan Bygde paper.