Improved/Fixed analysis that performs forward/backward refinement.
If the program has assertions we run first a forward analysis from the
initial states followed by a backward analysis from the error
states. We repeat this process until no more refinement is possible.

This refinement loop does not affect the invariants returned by the
analysis which are *always* the ones computed by the first forward
analysis. However, this analysis can potentially prove more
assertions: an assertion can be proven safe if starting from the
negation of the assertion condition (i.e. error states) we cannot find
an ascendant that is reachable from the initial states.

We collect all program assertions.

We compute dominator information because it can tell us if a block can
dominates another block where an assertion is defined.

We have modified the checkers API because the forward/backward
analysis can prove an assertion in cases where the forward propagation
of invariants on which Crab checkers rely on cannot. A common case is
when the assertion is a disequality that cannot be expressed by the
abstract domain.