FIX: overflow when zones uses int64 to represent DBM weights.
We should not call linear_constraint constructor using directly DBM
weights as one of the parameters. For instance, a linear constraint
less or equal can be built with two parameters: a variable (v) and int
number (k), representing v <= k. If the DBM weight is represented
using int64_t then it can overflow when casted to int (signed 32 bit
integers). The fix is to wrap the DBM weight in a number_t object. In
that way, a different constructor taking a variable v and number_t
constant is used so no casting happens.

We should probably restrict the API to build linear constraints to
avoid using directly int. This would limit its use but avoids nasty
surprises.