refactor(variable): only const name() method.
The reason why we had a non-const name() method is to have a way to
get a reference to the variable factory from a variable. This is
actually very necessary for domains that create dynamically variables
(e.g., array domains) because they need to have access to the variable
factory at anytime.

The long-term solution is to make sure that each abstract domain has
its own variable factory. For instance, we already do that in the
reference and term domains.

Meanwhile, the domain can use the const name() method and removes its
constness in order to use the variable factory. This is a HACK but it
should be fine for now. AFIK, C++ only guarantees shallow immutability
so as long as we know what we are doing we should not trigger
any undefined behavior.