refactor: only in-place array store
We extended the CFG and the abstract domains to allow having array
stores with old and new arrays. The idea was to allow model programs
in memory SSA form. This code is unused and only array smashing
implemented it. So for now, we get rid of it. We can take it back if
really needed.