feat(abstract_domain): remove static methods
Only abstract domains can call the default constructor of itself or
another abstract domain. The analyzers and fixpoint iterators do not
call a default constructor of an abstract domain anymore.

This is important because it allows us to use the type-erasure idiom
to create a generate abstract domain so hopefully we can reduce
compilation time.