perf(generic_abstract_domain): reduce compilation time.
New abstract domain called generic_abstract_domain that by using the
type-erasure idiom it hides the actual abstract domain. This should
help in reducing compilation time since we should have now only once
copy of the analyzers.

It is up to the client to use or not. If tests are enabled then it can
be used with cmake option -DCRAB_USE_GENERIC_WRAPPER. By default, it
is not used.

Apart from compilation time, no differences in analysis time should be
noticeable although we need a bit more testing.

In terms of precision, AFIK the only place where it could be less
precise is in the bottom-up inter-procedural analysis where some
conversion happens between the top-down and bottom-up abstract
domains.  If the two domains are actually the same and contains array,
disjunctive or non-linear constraints then the conversion might lose
precision because it cannot tell they are actually the same domain.