refactor: remove C++ memory_region type.
Before a memory region was just an identifier (memory_region type).
That was enough for intra-procedural analysis but it was a problem for
inter-procedural analysis.

Now, a variable can be either a scalar, array or a region. All the
abstract operations related to references and regions take as
parameter a variable of region type. As a result, we can pass region
variables to functions so the inter-procedural analyses can propagate
region's contents as they do for scalar or array variables.

We have also re-implemented some abstract operations in the reference
domain (e.g., gep_ref) and implemented some unsupported abstract
operations such as load_from_array or store_to_array.