refactor(region domain): improve join if a region is uninitialized.
At loop headers is common to have one region uninitialized from the
entry and one initialized from the loop back-edge. A standard join
will lose all the information inferred from the loop body. But because
the region is initially uninitialized we can improve the information
about it using the information collected from the loop body.

The propagation is non-relational by projecting on regions and meet
the result.

The same effect would be achieved if we would unrolled once the loop.