refactor(region): do not return bottom if ref is null.
Technically, it is fine to return bottom since reading or writing
from/to a null reference should not be allowed. But we prefer to
ignore the statement conservatively. We can still check for it in the
checking phase and get bottom if there is an assertion before the
load/store statements.