refactor(td-inter): more control about how to store invariants.
Storing invariants is expensive because of memory and also because we
optionally store context-insensitive invariants, used e.g., for
pretty-printing, which calls the abstract domain join.

For the top-down inter-procedural analysis, to completely avoid
storing invariants run clam.py without options:

--crab-do-not-store-invariants
--crab-do-not-print-invariants

For the intra-procedural analysis, invariants are always stored if
crab-check is not none.