refactor(separate_domain): move project to separate_domain.
The projection in the separate domain can be done in two different
ways: (1) create a new separate domain and copy the relevant entries
or (2) remove from the separate domain the irrelevant entries.

Which approach is more efficient depends on the load of the patricia
tree.  If the number of relevant entries is relatively small then (1)
is better, otherwise we might prefer (2).  We have a couple of
parameters to allow switching dynamically between (1) and (2) that
might need to be tuned.