feat(region): support free and new intrinsics
We have extended the domain by keeping track whether a region might
have been deallocated or not. A region might have been deallocated if
any of its memory objects might have been deallocated. Since the
region domain smashes all the memory objects, as soon as there is a
call to free to a reference within a region we assume that the whole
region might have been deallocated.

The new intrinsics are:

- free(rgn, ref)
- b:= is_unfreed_or_null(rgn, ref)
- unfreed_or_null(rgn, ref)
- nonnull(ref)

The intrinsics free will be replaced by a CFG statement in the future.
The intrinsics is_unfreed_or_null allows to ask if a pointer is not a
dangling pointer. The intrinsics unfreed_or_null expresses that a
pointer cannot be dangling. The last intrinsic is added for
convenience but it is not strictly necessary since the domain language
already allows constrains over reference variables so the property
nonnull or null can be already expressed.