feat(domains): add new select abstract operation.
Recall the semantics of select(lhs, cond, e1, e2) is

if(cond)
  lhs := e1;
else
  lhs := e2;

We have three different versions to deal with numbers, booleans, and
references.

In all cases, we were handling select statements in the abstract
transformer by calling standard abstract operations (add constraint,
assign, and join). This is simple and generic but it can be
inefficient because of the call to the join operator. Better leave it
to the abstract domain.