refactor(discrete_domain): default constructor returns empty set.
This is an exception because the default constructor of an abstract
domain typically returns top (set with all possible variables).
However, for the case of discrete_domain (which represents a
parametric set with lattice operations) returning top in the default
constructor is not useful at all because top is absorbent (i.e., any
operation on top it returns top). This allows us to remove the domain
flat_killgen_domain which purposes was just to return empty set in the
default constructor.