new(domain):  separate_discrete_domain.
This domain separate_discrete_domain<Key, Element> is very similar to
separate_domain<Key,Value> and thus, it also maps element from Key to
values. In this case, the domain is specialized for Value =
discrete_domain<Element>. The semantics is a bit different from
separate_domain because bottom in discrete_domain simply means empty
set. We want to allow in the environment keys whose values are empy
sets (bottom). separate_domain does not allow that because if the
value is bottom the whole environment becomes bottom.