feat(domains): reasoning about allocation sites.
We have defined a new type allocation_site which is basically a
numerical identifier.

We have modified the Crab language, in particular the make_ref
statement, to take an allocation as parameter. In that way, each call
to make_ref can be associated to an allocation site.

We have modified the abstract operation ref_make from
abstract_domain_api to take an allocation site as parameter.

We have modified the region_domain abstract domain to propagate
allocation sites across the program.

Finally, we have extended abstract_domain_api with a new specialized
api called abstract_domain_results_api which provides two methods:

- is_null_ref
- get_allocation_sites

This API allows queries to the abstract domain about whether a pointer
is null or not, allocation sites of a given pointer, etc. Other future
methods can be is_aligned, etc.