fix(region_domain): tracking deallocation in ref_gep.
We were too conservative in the analysis of ref_gep when one of the
regions is not in the union-find (represented by m_rgn_dealloc_dom)
because we were removing the other region (and its whole equivalence
class). However, that is not needed.

All regions are supposed to be initialized. Even if we do not know
whether a region contains an object that might be deallocated we still
keep that region in the union-find. However, Clam might miss the
initialization of a region if the region is never modified or
read. This is not consistent with the semantics of Crab and it needs
to be fixed.