feat(region): new region domain without ghost variables.
The use of ghost variables makes significantly slower
joins/meets/widenings/narrowings due to the need of renaming and
projection.

Abstract domains such array domains cannot avoid the use of ghost
variables since they need to model multiple array elements. However,
the region domain is a special case where we only create one ghost
variable per region variable. We have written a new region domain that
does not use ghost variables for regions since we already have a
variable for each region. However, this requires changes in the
abstract domains because we are passing variables of region types to
all the lower-level domains which are not supposed to get those.

We have created the notion of a _foreign_ type meaning that from an
abstract domain perspective a variable has a different type from its
actual type. This foreign type is fixed and it has been defined only
for the region domain in mind.  For instance, a region of integers
will be seen always as an integer by the array domains. A region of
references is also seen as an integer by any lower-level domain and so
on.

For now, we have created a branch because there are two issues that
need to be resolved: (1) the code of the new region domain needs to be
factorized, and (2) I am not convinced that having this notion of
foreign types is the way to go.