fix(cfg): compare callsite and function declarations.
Sometimes it is useful to know if the called function of a callsite
and a function declaration refer to the same function. We wrote a
wrapper that allows to do this.  This wrapper wouldn't be needed if a
callsite would maintain a pointer to the callee but we do not do that
right now because it would affect how clients build CFGs but this is
something that should do at some point.