fix(td-inter): use the right callgraph WTO.
For a given callgraph, we generate multiple WTOs (weak topological
orderings), one per callgraph entry. We use WTOs to known when we
should run the checker or when a fixpoint of a callgraph node should
be considered stabilized. Since we run one analysis per entry, we
should use only the callgraph WTO related to that entry.

Note that we could store only the current callgraph WTO in the global
datastructure. We still keep all for now.