fix(linear-constraints): not indexable anymore
Being indexable has been an issue because we were using a hash
function to generate the index which is supposed to be unique. Thus,
linear_constraint is not a subclass of indexable anymore. To allow
linear_constraint objects to be stored in set-like containers, we have
implemented the lexicographical_compare function.