feat(region-domain): support for unknown regions
This is only implemented for now in region_without_ghost_domain but it
should be ported to the other region domain variants.

An unknown region is a region for which we do not know its type at
translation time. The domain assigns a type the first time a write
happens to the unknown region. Then, it checks that all the memory
reads/writes are consistent with that type. If not, it ignores soundly
the memory read/write.