feat(region-domain): support for unknown regions.
region_with_is_deref_domain has been extended to be more precise on
unknown regions.

An unknown region is a region for which we do not know its type at
translation time. The domain assigns a type the first time a write
happens to the unknown region. Then, it checks that all the memory
reads/writes are consistent with that type. If not, it ignores soundly
the memory read/write.