refactor(domain): remove domain options as template parameters
In the past, abstract domain options have been set as template
parameters. This choice was made under the assumption that these
options were only tuned during the development of the abstract domain
and then stay fixed so users do not need to change them.

However, this is not the case sometimes and asking the user to
recompile the code it is really annoying. To solve this problem, we
have removed all template parameters that were used to represent
abstract domain options and instead, these options can be modified at
runtime via accessing to a global object
crab_domain_params_man.get(). Note that we are against the use of
global variables, however, here it seems the only option since we
don't have a abstract domain manager that is passed through the
analyses and the abstract domain operations.