feat(domains): new domain of uninterpreted functions
This domain maps program variables to uninterpreted functions.  The
code has been mostly extracted from term_equiv.hpp which implements
the term domain. The term domain combines uninterpreted functions with
a numerical domain. Thus, the uf_domain can be seen as a simplified
form of the term domain without the combination with a numerical
domain.