refactor(var-factory): create varnames from strings
New API that given same indexed_varname and a string it returns
always the same indexed_varname. This new method is useful for
abstract domains to create fresh varnames without requiring renaming.

TODO: the method should be marked as const but it would require to
make all the internal state mutable.