fix(varname_factory): dangling reference
After this fix, get_shadow_vars is an expensive operation. Instead of
allocating a new vector we could return an iterator that returns
references but I consider that it is not worth it for now.

This operation is intended to be used only for pretty-printing
purposes. AFIK, only Clam uses it and it is for making sure that it
does not print invariants expressed over non-LLVM Values.