refactor(domains): remove array graph domain
This was a proof-of-concept implementation of the paper A
Partial-Order Approach to Array Content Analysis. However, the
implementation is too inneficient and not well designed (e.g., use of
global state). Due to lack of users, it is not worth it to keep it.

We will commit soon a completely new implementation