fix(array-smashing): do not pass array variables to base domain.
Passing array variables to numerical domains is okay under the
assumption that the numerical domain does not really care about the
type of a variable. This is actually true in most cases at least for
numerical domains that reason about mathematical integers. However,
this is not true for domains that care, for instance, about integer
bitwidths. With this fix, the wrapped interval domain can also be
used.