refactor(region): decoupling all ghosting stuff
We have now the notion of a ghost variable manager to which all
abstract operations of the base domain are delegated to. The manager
knows how to create ghost variables and also whether renaming must be
done at binary operations. One clear advantage is that can switch
between ghosting naming schemes.