feat(domains): new state-partitioning disjunctive domain
The domain is called product_value_partitioning_domain.  For a given
variable, the domain keeps different abstract states for each possible
value of the variable. For efficiency, the values of a variable are
overapproximated by intervals.

{v1 == intv1 -> absState1, v1 == intv2 -> absState2, v2 == intv3 -> absState3, ...}

The domain understands two intrinsics value_partition_start and
value_partition_end. The former tells the domain that it should start
partitioning on a variable. The latter tells the domain to merge all
partitions for a given variable. In that way, the user can control the
size of the partitions.