feat(packing-domain): new numerical variable packing
The idea is to partition dynamically variables into disjoint packs
(aka buckets) and model each pack with a different abstract state.
This domain relies heavily on the union-find domain to represent each
pack as a different equivalence class. The analysis of a statement s
merges first all the equivalence classes associated to s variables and
then applies the corresponding transfer function.

This domain is intended to be used with domains such as octagons or
polyhedra.