fix: linear_constraint is always signed
linear_constraint has a field that indicated whether the linear
constraint should be interpreted as signed or unsigned. We removed
that field because Crab numerical domains do not support unsigned
constraints anyway. Moreover, linear_constraint had already signed
interpretation because depends on linear_expression so that was
inconsistent.

We have also fixed a bug in linear_constraint::is_false(). Thanks to
Enea Zaffanella