fix(bool-domain): propagation in assume(not b)
The kind of facts that we keep track in m_bool_to_lincsts (resp
m_bool_to_bools) is of the form:

  if b becomes true then constraint C (resp. boolean b') MUST be true

Thus, if b is false then nothing can really be said about C (resp b')
but we were propagating incorrectly that C (resp b') was false.

We added a test that passed incorrectly before (assertion was proven)
and now it doesn't pass anymore (assertion cannot be proven).