feat(domain): add a new abstract domain to infer object's properties
A new abstract domain followed by a new memory abstraction based on
DSA analysis to infer relational properties between fields. The domain
is split into several subdomains in order to reduce the complexity of
the anlysis. In terms of precision, the object domain maintain
equalities property between memory load and store to refine the relational
properties between fields. There also involves domain reduction for
precision refinement. Overall, This is a new abstract domain
for the refinement of DSA analysis to infer properties of memory objects.